+++ src/components/Persons/PersonEdit.tsx
@@ -1,42 +1,13 @@
-import React, { FormEvent, ReactElement, useState } from 'react';
+import React, { ReactElement } from 'react';
 import { useParams } from 'react-router';
-import { Redirect, useHistory, useLocation } from 'react-router-dom';
+import { Redirect } from 'react-router-dom';
 import { QueryRenderer } from 'react-relay';
 import environment from '../../relay-env';
 import graphql from 'babel-plugin-relay/macro';
-import { Button, Form, Spinner } from 'react-bootstrap';
 import { PersonEditQuery } from './__generated__/PersonEditQuery.graphql';
-import notifier from 'codex-notifier';
-import ContentWrapper from '../ContentWrapper';
-import commitMutation from 'relay-commit-mutation-promise';
-import {
-  PersonEditUpdateMutation,
-  PersonEditUpdateMutationResponse, UpdatePersonInput
-} from './__generated__/PersonEditUpdateMutation.graphql';
-import Input from '../utils/Input';
-import Textarea from '../utils/Textarea';
 import LoadingPlaceholder from '../utils/LoadingPlaceholder';
-import { LabeledArrayOfInputs } from '../utils/ArrayOfInputs';
-
-/**
- * Executes update mutation for person
- *
- * @param input - updated person object
- */
-export function update(input: UpdatePersonInput): Promise<PersonEditUpdateMutationResponse> {
-  return commitMutation<PersonEditUpdateMutation>(environment, {
-    mutation: graphql`
-      mutation PersonEditUpdateMutation($input: UpdatePersonInput!) {
-        person {
-          update(input: $input) {
-            recordId
-          }
-        }
-      }
-    `,
-    variables: { input },
-  });
-}
+import PersonEditForm from './PersonEditForm';
+import notifier from 'codex-notifier';
 
 /**
  * Page with form for person editing
@@ -44,76 +15,13 @@ export function update(input: UpdatePersonInput): Promise<PersonEditUpdateMutati
 export default function PersonEdit(): ReactElement {
   const { id } = useParams();
 
-  const [input, setInput] = useState<UpdatePersonInput | null>(null);
-  const [isLoading, setLoadingStatus] = useState(false);
-
-  const history = useHistory();
-  const location = useLocation();
-
-  /**
-   * Push location back to entity view page
-   */
-  const pushLocationBack = (): void => {
-    const entityListPath = location.pathname.replace('/edit', '');
-
-    history.push(entityListPath);
-  };
-
-  /**
-   * Saves updated person to API
-   *
-   * @param e - submit form event
-   */
-  const updatePerson = async (e: FormEvent): Promise<void> => {
-    e.preventDefault();
-    if (!input) {
-      notifier.show({
-        message: 'Please make changes in person',
-        style: 'error',
-        time: 5000,
-      });
-
-      return;
-    }
-
-    setLoadingStatus(true);
-    if ('id' in input) {
-      try {
-        await update(input);
-        notifier.show({
-          message: 'Successfully updated',
-          style: 'success',
-          time: 5000,
-        });
-        setLoadingStatus(false);
-        pushLocationBack();
-      } catch {
-        setLoadingStatus(false);
-        notifier.show({
-          message: 'Something went wrong',
-          style: 'error',
-          time: 5000,
-        });
-      }
-    }
-  };
-
   return (
     <QueryRenderer<PersonEditQuery>
       environment={environment}
       query={graphql`
         query PersonEditQuery($id: GlobalId!) {
           person(id: $id) {
-            id
-            lastName
-            firstName
-            patronymic
-            pseudonym
-            professions
-            description
-            birthDate
-            deathDate
-            wikiLink
+            ...PersonEditForm_originalPerson
           }
         }
       `}
@@ -138,123 +46,7 @@ export default function PersonEdit(): ReactElement {
           return <Redirect to='/persons'/>;
         }
 
-        return (
-          <ContentWrapper>
-            <Form onSubmit={updatePerson}>
-              <Input
-                label='Last name'
-                onChange={value => setInput({
-                  ...input,
-                  id,
-                  lastName: value,
-                })}
-                required
-                value={input?.lastName || props.person.lastName}
-              />
-              <Input
-                label='First name'
-                onChange={value => setInput({
-                  ...input,
-                  id,
-                  firstName: value,
-                })}
-                required
-                value={input?.firstName || props.person.firstName}
-              />
-              <Input
-                label='Patronymic'
-                onChange={value => setInput({
-                  ...input,
-                  id,
-                  patronymic: value,
-                })}
-                value={input?.patronymic || props.person.patronymic}
-              />
-              <Input
-                label='Pseudonym'
-                onChange={value => setInput({
-                  ...input,
-                  id,
-                  pseudonym: value,
-                })}
-                value={input?.pseudonym || props.person.pseudonym}
-              />
-              <LabeledArrayOfInputs
-                addButtonText='Add profession...'
-                label='Professions'
-                onChange={value => setInput({
-                  ...input,
-                  id,
-                  professions: value,
-                })}
-                removeButtonText='Remove profession'
-                /**
-                 * Converts ReadonlyArray to simple array
-                 */
-                value={input?.professions || props.person.professions.concat()}
-              />
-              <Textarea
-                label='Description'
-                onChange={value => setInput({
-                  ...input,
-                  id,
-                  description: value,
-                })}
-                value={input?.description || props.person.description}
-              />
-              <Input
-                label='Birth date'
-                onChange={value => setInput({
-                  ...input,
-                  id,
-                  birthDate: value,
-                })}
-                value={input?.birthDate || props.person.birthDate}
-              />
-              <Input
-                label='Death date'
-                onChange={value => setInput({
-                  ...input,
-                  id,
-                  deathDate: value,
-                })}
-                value={input?.deathDate || props.person.deathDate}
-              />
-              <Input
-                label='Wiki link'
-                onChange={value => setInput({
-                  ...input,
-                  id,
-                  wikiLink: value,
-                })}
-                value={input?.wikiLink || props.person.wikiLink}
-              />
-              <div>
-                <Button
-                  className='m-1'
-                  type='submit'
-                >
-                  {isLoading ? (
-                    <Spinner
-                      animation='border'
-                      aria-hidden='true'
-                      as='span'
-                      role='status'
-                      size='sm'
-                    />
-                  ) : ('Save')}
-                </Button>
-                <Button
-                  className='m-1'
-                  onClick={() => pushLocationBack()}
-                  variant='outline-danger'
-                >
-                  Cancel
-                </Button>
-              </div>
-            </Form>
-          </ContentWrapper>
-        );
+        return <PersonEditForm originalPerson={props.person}/>;
       }}
       variables={{ id }}
     />
